llm/book-ollama-opensourcellm-aiagent-develop-entry/ch06/02_langchain_hub/pyproject.toml
[project]
name = "ollama-mcp-tutorial"
version = "0.1.0"
description = "간단한 Ollama 챗봇 예제"
requires-python = ">=3.10"
dependencies = [
    "langchain-core==0.3.61",
    "langchain-ollama==0.3.3",
    "langchain==0.3.25",
    "beautifulsoup4",
    "python-dotenv",
    "langchain-community"
]